Change the default jenkins parameters
Change to http and sanbase so the cronjob will use them and not get rate
limited.
Change to 180 seconds as it's also enough - longer tests can be ran from
jenkins. Main reason for change is that every brahch/PR runs all tests
and can take a lot of time to build